`Paywalls`: extract `TemplatePackageSetting` and add `TemplateTierSetting` (#3663)
This will be used for multi-tier paywalls.
