+++ RevenueCatUI/Templates/TemplateViewType.swift
@@ -60,21 +60,6 @@ extension TemplateViewType {
 }
 
 @available(iOS 15.0, macOS 12.0, tvOS 15.0, watchOS 8.0, *)
-extension PaywallTemplate {
-
-    var packageSetting: TemplateViewConfiguration.PackageSetting {
-        switch self {
-        case .template1: return .single
-        case .template2: return .multiple
-        case .template3: return .single
-        case .template4: return .multiple
-        case .template5: return .multiple
-        }
-    }
-
-}
-
-@available(iOS 15.0, macOS 12.0, tvOS 15.0, watchOS 8.0, *)
 @available(macOS, unavailable)
 @available(tvOS, unavailable)
 extension PaywallData {
